Move auctions panel into v1 folder
