+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/v1/adunitPanel/index.tsx
@@ -23,7 +23,10 @@ import { useTabs } from '@google-psat/design-system';
 /**
  * Internal dependencies
  */
-import { useCookie, useProtectedAudience } from '../../../../../stateProviders';
+import {
+  useCookie,
+  useProtectedAudience,
+} from '../../../../../../stateProviders';
 import Panel from './panel';
 
 interface AdunitPanelProps {
