+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/v1/adunitPanel/panel.tsx
@@ -29,7 +29,7 @@ import type {
  * Internal dependencies
  */
 import Tile from './tile';
-import type { AdUnitLiteral } from '../../explorableExplanation';
+import type { AdUnitLiteral } from '../../../explorableExplanation';
 import MatrixHOC from './matrixContainer';
 
 interface PanelProps {
