+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/v1/table/bottomTray.tsx
