+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/index.tsx
@@ -27,7 +27,7 @@ import {
  * Internal dependencies.
  */
 import InterestGroups from './interestGroups';
-import Auctions from './auctions';
+import Auctions from './auctions/v1';
 import Bids from './bids';
 import Panel from './panel';
 import AdUnits from './adUnits';
